Include thanks to in-kind sponsors
