+++ src/components/SponsorIcons/SponsorIcons.module.scss
@@ -1,0 +1,18 @@
+.logosList {
+  list-style: none;
+  display: flex;
+  align-content: center;
+  align-items: center;
+  //justify-content: center;
+  gap: 12px;
+  max-width: 100%;
+  overflow-x: hidden;
+  padding: 0;
+  margin: 0;
+  color: var(--ifm-font-color-base);
+
+  img {
+    max-width: 30px;
+    max-height: 30px;
+  }
+}
