+++ src/components/SponsorNames/SponsorNames.module.scss
@@ -1,0 +1,13 @@
+.list {
+  list-style: none;
+  padding: 0;
+  margin: 0;
+}
+
+.sponsor {
+  color: inherit;
+}
+
+.sponsorLogo {
+  border-radius: 0;
+}
