+++ src/components/SponsorTiers/index.ts
@@ -1,0 +1,1 @@
+export * from './SponsorTiers'
