--- src/components/UsedBy/UsedBy.module.scss
@@ -1,31 +1,0 @@
-.strapLine {
-  text-align: center;
-}
-
-.logosList {
-  list-style: none;
-  display: flex;
-  align-content: center;
-  align-items: center;
-  justify-content: center;
-  gap: 24px;
-  max-width: 100%;
-  overflow-x: hidden;
-  color: var(--ifm-font-color-base);
-
-  img {
-    max-width: 120px;
-    max-height: 50px;
-    filter: #{'grayscale()'};
-    opacity: 0.5;
-
-    [data-theme='dark'] & {
-      filter: #{'grayscale() invert()'};
-    }
-
-    @media all and (max-width: 600px) {
-      max-width: 80px;
-      max-height: 30px;
-    }
-  }
-}
