--- src/components/UsedBy/index.ts
@@ -1,1 +1,0 @@
-export * from './UsedBy'
