+++ src/data/in-kind-sponsors.json
@@ -1,0 +1,27 @@
+[
+  {
+    "name": "1Password",
+    "logo": "/img/sponsors/1password.svg",
+    "url": "https://1password.com"
+  },
+  {
+    "name": "Algolia",
+    "logo": "/img/sponsors/algolia.svg",
+    "url": "https://www.algolia.com"
+  },
+  {
+    "name": "Buttondown",
+    "logo": "/img/sponsors/buttondown.svg",
+    "url": "https://buttondown.com"
+  },
+  {
+    "name": "GitHub",
+    "logo": "/img/sponsors/github.svg",
+    "url": "https://github.com"
+  },
+  {
+    "name": "Netlify",
+    "logo": "/img/sponsors/netlify.svg",
+    "url": "https://www.netlify.com"
+  }
+]
