+++ src/pages/sponsors.module.scss
@@ -1,54 +1,4 @@
-.list {
-  list-style: none;
-  padding: 0;
-  margin: 0;
-}
-
-.tier {
-  width: 20%;
-}
-
-.card {
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-  box-shadow: 0 1.5px 3px 0 rgb(0 0 0 / 15%);
-  border: 1px solid var(--ifm-color-emphasis-200);
-}
-
-.amount {
-  display: block;
-  font-weight: 300;
-  font-size: 2.5rem;
-  line-height: 1em;
-}
-
-.frequency {
-  display: block;
-  color: var(--ifm-color-secondary-darkest);
-  font-size: 0.875rem;
-}
-
-.blurb {
-  flex: 1;
-}
-
-// breakpoint for infima grid dropping down to single column
-@media all and (max-width: 996px) {
-  .tier {
-    width: 100%;
-  }
-}
-
 .reasonIllustration {
   max-width: 80%;
   max-height: 150px;
-}
-
-.sponsor {
-  color: inherit;
-}
-
-.sponsorLogo {
-  border-radius: 0;
 }
