+++ src/pages/sponsors.tsx
@@ -1,4 +1,4 @@
-import React, { FC } from 'react'
+import React from 'react'
 import Layout from '@theme/Layout'
 import Link from '@docusaurus/Link'
 import styles from './sponsors.module.scss'
@@ -8,118 +8,11 @@ import { Hero } from '@site/src/components/Hero'
 import goldSponsors from '@site/src/data/gold-sponsors.json'
 import silverSponsors from '@site/src/data/silver-sponsors.json'
 import bronzeSponsors from '@site/src/data/bronze-sponsors.json'
+import inKindSponsors from '@site/src/data/in-kind-sponsors.json'
 import { DownloadCounter } from '@site/src/components/DownloadCounter'
-
-interface Sponsor {
-  name: string
-  logo: string
-  url: string
-  blurb?: string
-}
-
-const numberFormat = new Intl.NumberFormat('en', {
-  style: 'currency',
-  currency: 'USD',
-  maximumFractionDigits: 0,
-})
-
-const tiers = [
-  {
-    title: 'Individual',
-    description: 'For individuals who want to support the project',
-    amount: 5,
-    url: 'https://opencollective.com/cucumber/contribute/monthly-backers-182/checkout',
-    from: true,
-  },
-  {
-    title: 'Bronze',
-    description:
-      'Support the contributors with a monthly donation and, if you want, get your name and a small logo on the Cucumber website with a link to your site.',
-    amount: 100,
-    url: 'https://opencollective.com/cucumber/contribute/bronze-sponsors-181/checkout',
-    from: false,
-  },
-  {
-    title: 'Silver',
-    description:
-      'Support the contributors with a monthly donation and, if you want, get your name and a medium logo on the Cucumber website with a link to your site.',
-    amount: 500,
-    url: 'https://opencollective.com/cucumber/contribute/gold-sponsors-3224/checkout',
-    from: false,
-  },
-  {
-    title: 'Gold',
-    description:
-      'Support the contributors with a monthly donation and, if you want, get your name, large logo, and subtitle on the Cucumber website with a link to your site.',
-    amount: 1000,
-    url: 'https://opencollective.com/cucumber/contribute/gold-82673/checkout',
-    from: false,
-  },
-  {
-    title: 'Platinum',
-    description:
-      'Support the contributors with a monthly donation and, if you want, get your name, large logo, subtitle and short paragraph on the Cucumber website with a link to your site.',
-    amount: 2500,
-    url: 'https://opencollective.com/cucumber/contribute/platinum-82674/checkout',
-    from: false,
-  },
-] as const
-
-const Tiers: FC = () => {
-  return (
-    <ol className={clsx('row', styles.list)}>
-      {tiers.map((tier) => {
-        return (
-          <li key={tier.amount} className={clsx('col margin-bottom--lg', styles.tier)}>
-            <div className={clsx('card', styles.card)}>
-              <div className="card__header">
-                <h3>{tier.title}</h3>
-              </div>
-              <div className="card__body">
-                <b className={styles.amount}>{numberFormat.format(tier.amount)}</b>
-                <span className={styles.frequency}>{tier.from && 'or above, '}monthly</span>
-                <p className={clsx('margin-vert--md', styles.blurb)}>{tier.description}</p>
-              </div>
-              <div className="card__footer">
-                <Link className="button button--block button--primary" href={tier.url}>
-                  Sponsor
-                </Link>
-              </div>
-            </div>
-          </li>
-        )
-      })}
-    </ol>
-  )
-}
-
-interface SponsorsListProps {
-  sponsors: Sponsor[]
-  size?: 'sm' | 'md' | 'lg' | 'xl'
-}
-
-const SponsorsList: FC<SponsorsListProps> = ({ sponsors, size = 'lg' }) => {
-  if (sponsors.length === 0) return null
-  return (
-    <ul className={styles.list}>
-      {sponsors.map((sponsor) => (
-        <li key={sponsor.name} className="margin-bottom--md">
-          <Link href={sponsor.url} rel="noopener noreferrer" className={`avatar ${styles.sponsor}`}>
-            <img
-              className={`avatar__photo avatar__photo--${size} ${styles.sponsorLogo}`}
-              src={sponsor.logo}
-              alt={`${sponsor.name} logo`}
-            />
-            <div className="avatar__intro">
-              <div className="avatar__name">{sponsor.name}</div>
-              {sponsor.blurb && <small className="avatar__subtitle">{sponsor.blurb}</small>}
-            </div>
-          </Link>
-        </li>
-      ))}
-    </ul>
-  )
-}
+import { SponsorIcons } from '@site/src/components/SponsorIcons'
+import { SponsorNames } from '@site/src/components/SponsorNames'
+import { SponsorTiers } from '@site/src/components/SponsorTiers'
 
 export default function Sponsors() {
   return (
@@ -168,19 +61,26 @@ export default function Sponsors() {
               {goldSponsors.length > 0 && (
                 <>
                   <h3>Gold Sponsors</h3>
-                  <SponsorsList sponsors={goldSponsors} size="xl" />
+                  <SponsorNames sponsors={goldSponsors} size="xl" />
                 </>
               )}
               {silverSponsors.length > 0 && (
                 <>
                   <h3>Silver Sponsors</h3>
-                  <SponsorsList sponsors={silverSponsors} size="lg" />
+                  <SponsorNames sponsors={silverSponsors} size="lg" />
                 </>
               )}
               {bronzeSponsors.length > 0 && (
                 <>
                   <h3>Bronze Sponsors</h3>
-                  <SponsorsList sponsors={bronzeSponsors} size="sm" />
+                  <SponsorNames sponsors={bronzeSponsors} size="sm" />
+                </>
+              )}
+              {inKindSponsors.length > 0 && (
+                <>
+                  <h3>In-Kind Sponsors</h3>
+                  <p>Our thanks to these companies for supporting open source with free/subsidised plans.</p>
+                  <SponsorIcons sponsors={inKindSponsors} />
                 </>
               )}
             </div>
@@ -191,7 +91,7 @@ export default function Sponsors() {
           <p className="text--center">
             If you can, we'd love for you to commit a regular monthly amount to support Cucumber.
           </p>
-          <Tiers />
+          <SponsorTiers />
           <div className="text--center">
             <p className="margin-bottom--sm">
               Or, you can always make a{' '}
